RavenDB-5177 Abstract RavenThreadPool and allow custom implementations
