+++ Raven.Tests.Issues/Raven.Tests.Issues.csproj
@@ -151,6 +151,7 @@
     <Compile Include="RavenDB-4705.cs" />
     <Compile Include="RavenDB-4904.cs" />
     <Compile Include="RavenDB-5151.cs" />
+    <Compile Include="RavenDB-5177.cs" />
     <Compile Include="RavenDB_4645.cs" />
     <Compile Include="RavenDB_4918.cs" />
     <Compile Include="RavenDb_4706.cs" />
